💄[style] 구성원 목록 스타일 수정
